Add "surroundings" and "positions" environments
The former observes whether there is danger in the 8 tiles immediately
surrounding the snake's head as well as whether there is food in
any/multiple of the 4 cardinal directions relative to the head.

The latter observes the distance of the nearest danger in each
of the 4 directions as well as the relative position of the food
relative to the head.

Each environment needs its own training script. I've started working on
breaking out the training portion into something a bit more modular, but
there will need to be a lot of modification between each one, so not
going to overdo it.